Reorganize with provider in Root component
