+++ src/containers/rootReducer.js
@@ -1,0 +1,8 @@
+const initialState = {}
+
+export function rootReducer (state=initialState, action) {
+  switch (action.type) {
+    default:
+      return state
+  }
+}
