+++ src/utilities/registerServiceWorker.js
